+++ src/examples/uniprot/index.ts
@@ -1,11 +1,10 @@
-import './index.html';
 
 import {RcsbFv3DUniprot} from "../../RcsbFv3D/RcsbFv3DUniprot";
 
 document.addEventListener("DOMContentLoaded", function(event) {
 
     const panel3d = new RcsbFv3DUniprot({
-        elementId:"pfv",
+        elementId:"null",
         config:{
             upAcc:"P01112"
         }